requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-datatables==0.7.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
h11==0.16.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
uvicorn==0.34.2
